Revert to the previous behavior of DeviceId >= 16.
Some users are reporting performance issues with the input refactor,
which might be caused by the change in behavior when DeviceId >= 16,
which went from using the first successful device, to scanning all of
the devices and merging inputs from all of them. It's possible this is
causing multiple device scans per frame, so revert back to the previous
behavior.

Make the previously new behavior accessible by specifying a negative
DeviceId.